refactor: extract widgets and reader from KCPPS dialog and GGUF parser
- Extract GGUFFileReader (shared binary reader) from gguf_parser.dart
- Extract GGUFModelInfo into own file, re-export from gguf_parser.dart
- Extract VramEstimateBreakdown typedef into vram_estimator.dart
- Extract VramUsageSection, ContextManagementSelector, GpuInfoTile widgets
- Wire 3 new widgets into generate_kcpps_dialog.dart, remove _buildVramSection
- Fix Colors.blueAccent -> colorScheme.primary and Colors.redAccent -> colorScheme.error
- Remove dead ByteData.sublistView from gguf_parser.dart (no import needed)
- All 37 tests pass, flutter analyze shows 0 issues
